WIZ-587 fix from review
