+++ tests/Company/CompanyServiceTest/testRegisterACompanyC2CWithIbanAndBicAndAddress.yml
@@ -7,7 +7,7 @@
             Host: wizaplace.loc
             Accept-Encoding: null
             Authorization: 'Basic dXNlckB3aXphcGxhY2UuY29tOnBhc3N3b3Jk'
-            User-Agent: Wizaplace-PHP-SDK/dev-hotfix/wiz-587-smoney@a817111
+            User-Agent: Wizaplace-PHP-SDK/dev-hotfix/wiz-587-smoney@5a80f68
             VCR-index: '0'
             Accept: null
     response:
@@ -22,10 +22,10 @@
             Connection: keep-alive
             X-Powered-By: PHP/7.2.14
             Cache-Control: 'no-cache, private'
-            Date: 'Tue, 05 Feb 2019 07:30:28 GMT'
+            Date: 'Wed, 06 Feb 2019 09:00:40 GMT'
             Content-Language: fr
-            X-Debug-Token: e27ba8
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/e27ba8'
+            X-Debug-Token: 87ac80
+            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/87ac80'
         body: '{"id":3,"apiKey":"1AAegjK0mV9MVTYeV1chpp5Yvp1UvLBTGwGwMyxE"}'
 -
     request:
@@ -36,11 +36,11 @@
             Expect: null
             Accept-Encoding: null
             Content-Type: application/json
-            User-Agent: Wizaplace-PHP-SDK/dev-hotfix/wiz-587-smoney@a817111
+            User-Agent: Wizaplace-PHP-SDK/dev-hotfix/wiz-587-smoney@5a80f68
             Authorization: 'token 1AAegjK0mV9MVTYeV1chpp5Yvp1UvLBTGwGwMyxE'
             VCR-index: '1'
             Accept: null
-        body: '{"name":"Super nom","email":"user@wizaplace.com","description":null,"slug":null,"address":"address","country":"FR","zipcode":"zipCode","city":"city","phoneNumber":null,"url":null,"fax":null,"vatNumber":null,"siretNumber":null,"rcs":null,"legalStatus":null,"capital":null,"iban":"AD1200012030200359100100","bic":"AGFBFRCC","extra":[]}'
+        body: '{"name":"Super nom","email":null,"description":null,"slug":null,"address":"address","country":"FR","zipcode":"zipCode","city":"city","phoneNumber":null,"url":null,"fax":null,"vatNumber":null,"siretNumber":null,"rcs":null,"legalStatus":null,"capital":null,"iban":"AD1200012030200359100100","bic":"AGFBFRCC","extra":[]}'
     response:
         status:
             http_version: '1.1'
@@ -53,9 +53,9 @@
             Connection: keep-alive
             X-Powered-By: PHP/7.2.14
             Cache-Control: 'no-cache, private'
-            Date: 'Tue, 05 Feb 2019 07:30:29 GMT'
+            Date: 'Wed, 06 Feb 2019 09:00:42 GMT'
             Content-Language: fr
-            X-Debug-Token: 547b07
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/547b07'
-            Set-Cookie: 'sf_redirect=%7B%22token%22%3A%22547b07%22%2C%22route%22%3A%22api_register_c2c_company%22%2C%22method%22%3A%22POST%22%2C%22controller%22%3A%7B%22class%22%3A%22Wizacha%5C%5CAppBundle%5C%5CController%5C%5CApi%5C%5CCompanyController%22%2C%22method%22%3A%22registerC2CAction%22%2C%22file%22%3A%22%5C%2Fvar%5C%2Fwww%5C%2Fhtml%5C%2Fsrc%5C%2FAppBundle%5C%2FController%5C%2FApi%5C%2FCompanyController.php%22%2C%22line%22%3A70%7D%2C%22status_code%22%3A201%2C%22status_text%22%3A%22Created%22%7D; path=/; httponly'
-        body: '{"id":8,"name":"Super nom","slug":"super-nom-2","email":"user@wizaplace.com","description":"","zipcode":"zipCode","address":"address","city":"city","state":"","country":"FR","phoneNumber":"","legalStatus":"Vendeur particulier","siretNumber":"","vatNumber":"","capital":"","rcs":"","fax":"","url":"","iban":"AD1200012030200359100100","bic":"AGFBFRCC","legalRepresentativeFirstName":"","legalRepresentativeLastName":"","extra":[]}'
+            X-Debug-Token: e307c8
+            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/e307c8'
+            Set-Cookie: 'sf_redirect=%7B%22token%22%3A%22e307c8%22%2C%22route%22%3A%22api_register_c2c_company%22%2C%22method%22%3A%22POST%22%2C%22controller%22%3A%7B%22class%22%3A%22Wizacha%5C%5CAppBundle%5C%5CController%5C%5CApi%5C%5CCompanyController%22%2C%22method%22%3A%22registerC2CAction%22%2C%22file%22%3A%22%5C%2Fvar%5C%2Fwww%5C%2Fhtml%5C%2Fsrc%5C%2FAppBundle%5C%2FController%5C%2FApi%5C%2FCompanyController.php%22%2C%22line%22%3A70%7D%2C%22status_code%22%3A201%2C%22status_text%22%3A%22Created%22%7D; path=/; httponly'
+        body: '{"id":9,"name":"Super nom","slug":"super-nom-3","email":"user@wizaplace.com","description":"","zipcode":"zipCode","address":"address","city":"city","state":"","country":"FR","phoneNumber":"","legalStatus":"Vendeur particulier","siretNumber":"","vatNumber":"","capital":"","rcs":"","fax":"","url":"","iban":"AD1200012030200359100100","bic":"AGFBFRCC","legalRepresentativeFirstName":"","legalRepresentativeLastName":"","extra":[]}'
